+++ flockwise/models/employee.js
@@ -2,6 +2,7 @@
 
 import { Schema, model, models } from 'mongoose';
 import Task from './task'; // Import the Task model
+import Paystub from './paystub'; // Import the Paystub model
 
 const employeeSchema = new Schema({
   employeeID: {
@@ -25,6 +26,10 @@ const employeeSchema = new Schema({
   //   type: [Benefits.schema], // make another object for benefits or implement however desired
   //   ref: 'BenefitsManagement',
   // },
+  paystubs: {
+    type: [Paystub.schema],
+    default: [],
+  },
   salary: {
     type: Number,
     default: 7.25,
